chore: move doc generation to ts
