+++ package.json
@@ -21,9 +21,7 @@
     "url": "https://github.com/maplibre/maplibre-react-native"
   },
   "scripts": {
-    "generate": "yarn generate:fetch-style-spec && yarn generate:docs",
-    "generate:fetch-style-spec": "./scripts/download-style-spec.sh",
-    "generate:docs": "yarn node ./scripts/generate-docs",
+    "generate": "tsx scripts/generate-docs.ts",
     "test": "jest",
     "lint": "yarn lint:eslint && yarn lint:tsc",
     "lint:eslint": "eslint .",
@@ -64,16 +62,18 @@
     "@babel/plugin-proposal-class-properties": "7.18.6",
     "@babel/runtime": "7.17.2",
     "@expo/config-plugins": "^7.2.5",
+    "@maplibre/maplibre-gl-style-spec": "21.0.0",
     "@react-native/babel-preset": "^0.74.88",
     "@react-native/metro-config": "^0.74.88",
     "@sinonjs/fake-timers": "^11.2.2",
     "@testing-library/react-native": "^12.4.3",
     "@tsconfig/node14": "^14.1.0",
+    "@types/ejs": "^3.1.5",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "babel-jest": "^29.6.0",
     "documentation": "^14.0.0",
-    "ejs": "^3.1.3",
+    "ejs": "^3.1.10",
     "ejs-lint": "^2.0.0",
     "eslint": "^8.57.1",
     "eslint-config-universe": "13.0.0",
@@ -86,9 +86,10 @@
     "pinst": "^3.0.0",
     "prettier": "3.3.3",
     "react": "18.2.0",
-    "react-docgen": "rnmapbox/react-docgen#rnmapbox-dist",
+    "react-docgen": "^7.1.0",
     "react-native": "^0.74.6",
     "react-test-renderer": "18.2.0",
+    "tsx": "^4.19.2",
     "typescript": "^5.3.3"
   },
   "resolutions": {
